Upgrade Strapi to 4.25.13
