+++ package.json
@@ -11,11 +11,14 @@
   },
   "devDependencies": {},
   "dependencies": {
-    "@strapi/strapi": "4.14.6",
-    "@strapi/plugin-users-permissions": "4.14.6",
-    "@strapi/plugin-i18n": "4.14.6",
+    "@strapi/strapi": "4.25.13",
+    "@strapi/plugin-users-permissions": "4.25.13",
+    "@strapi/plugin-i18n": "4.25.13",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^5.2.0",
+    "styled-components": "5.2.1",
     "pg": "8.8.0",
-    "@notum-cz/strapi-plugin-content-versioning": "^0.4.7",
     "node-fetch": "2"
   },
   "author": {
